requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==21.2.0
azure-core==1.32.0
azure-identity==1.20.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brotli==1.0.9
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
cryptography==44.0.1
cycler==0.11.0
decorator==4.4.2
defusedxml==0.8.0rc2
distlib==0.3.9
distro==1.7.0
distro-info
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml
lz4==3.1.2
matplotlib>=3.5.0
matplotlib-inline==0.1.3
more-itertools==8.10.0
mpmath==1.2.1
msal==1.31.1
msal-extensions==1.2.0
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
olefile==0.46
packaging==24.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.3.6
ply==3.11
portalocker==2.10.1
prompt-toolkit==3.0.28
psycopg2>=2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
Pygments==2.11.2
PyJWT==2.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==1.0.1
python3-openid==3.2.0
pythran==0.10.0
pytz==2022.1
PyYAML>=5.0.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy>=1.9.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.3.1
sqlparse==0.5.3
sympy==1.9
traitlets==5.1.1
typing_extensions==4.12.2
ufoLib2==0.13.1
urllib3==2.3.0
virtualenv==20.29.1
wadllib>=1.0.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==1.0.0
whitenoise==6.0.0
drf-nested-routers>=0.92.1
texlive-fonts-extra

# Install Python first
# requirements.txt store all the need packages
# Run "pip install -r requirements.txt"
